Bit by interaction in Function name and Unicode...
If an operator name like "Function" has unicode associated with it, then
the YAML key *has* to be the operator name.

This is because we are doing weird Function-name to unicode translation
in *input*

If there is an ASCII version of it like "&" for the function, then
the YAML name doesn't have to match (since there is no auto unicode
translation going on).